+++ Sources/UI/DynamicView/DynamicViewModel.swift
@@ -94,6 +94,10 @@ struct WidgetWrapper: Decodable {
             value = try WidgetVersion(from: decoder)
         case .customerCard:
             value = try WidgetCustomerCard(from: decoder)
-        }
+        case .developerMode:
+            value = try WidgetDeveloperMode(from: decoder)
+        case .multiValue:
+            value = try WidgetMultiValue(from: decoder)
+       }
     }
 }