+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unreleased]
 
 ### Added
+* DeveloperMode with 5 tabs on profile version field
+* Default developer password is "Snabble" customizable via AssetProviding key "SnabbleDeveloperPassword"
+* CheckIn / CheckOut for selected shops in DeveloperMode
+* Last Purchases and Receipts available for DynamicViews (eg. Dashboard / Profile Tab in SampleApp)
+* New WidgetMultiValue and corresponding SwiftUI view to display/store UserDefault Values
 
 ### Removed
 * AutoLayout-Helper Dependencies - Moved function in the SDK
@@ -14,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Update device location to search for shops
 
 ### Changed
+* updated certificates
 * renamed ProductProvider to ProductStore
 * renamed ProductStore functions by removing unnamed (underscore) arguments
 * Information Widget takes full width